refactor(value): 좌표 제한 검증 로직을 이동
